requirements.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
psycopg2-binary==2.9.*
PyMySQL==1.1.*
django-environ==0.11.*
django-cors-headers==4.3.*
django-filter==23.5
Pillow==10.*
WeasyPrint==62.*
celery==5.3.*
redis==5.0.*
